requirements-train.txt
pytorch_lightning
jsonargparse[signatures]
transformers==4.36.0
matplotlib
torchcrepe
pesq
fairseq==0.12.2
encodec
tensorboard
numpy==1.26.4